Resurrect 32 bit value hack
It feels dirty to submit this changelist. But unfortunately it's a lot quicker to resolve the underlying problem this way than to try and find the root cause for this manual offset adjustment. Especially since that hack existed since the beginning of this module, and many systems were built on top of that since then.